new project structure
to avoid to create new projects and repositories for everything I do, lets see if a multi-project is a solution. There will be still separate projects for everything, but the setup should be easier and it should also be a bit nicer to separate things instead of just doing that on java package level.